Move default config and config validation to core module
