+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/config/ConfigValidators.kt
@@ -1,12 +1,12 @@
-package io.gitlab.arturbosch.detekt.cli.config
+package io.gitlab.arturbosch.detekt.core.config
 
 import io.gitlab.arturbosch.detekt.api.ConfigValidator
 import io.gitlab.arturbosch.detekt.api.Notification
-import io.gitlab.arturbosch.detekt.core.reporting.red
 import io.gitlab.arturbosch.detekt.core.ProcessingSettings
+import io.gitlab.arturbosch.detekt.core.reporting.red
 import java.util.ServiceLoader
 
-internal fun checkConfiguration(settings: ProcessingSettings) {
+fun checkConfiguration(settings: ProcessingSettings) {
     val props = settings.config.subConfig("config")
     val shouldValidate = props.valueOrDefault("validation", true)
