+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/config/DefaultPropertiesConfigValidator.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.cli.config
+package io.gitlab.arturbosch.detekt.core.config
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.ConfigValidator
@@ -7,7 +7,6 @@ import io.gitlab.arturbosch.detekt.api.internal.CommaSeparatedPattern
 import io.gitlab.arturbosch.detekt.api.internal.DEFAULT_PROPERTY_EXCLUDES
 import io.gitlab.arturbosch.detekt.api.internal.DefaultRuleSetProvider
 import io.gitlab.arturbosch.detekt.api.internal.validateConfig
-import io.gitlab.arturbosch.detekt.cli.loadDefaultConfig
 import io.gitlab.arturbosch.detekt.core.ProcessingSettings
 import io.gitlab.arturbosch.detekt.core.RuleSetLocator
 
@@ -24,6 +23,6 @@ class DefaultPropertiesConfigValidator(
             val allExcludes = "$configExcludes,$DEFAULT_PROPERTY_EXCLUDES,$pluginExcludes"
             return CommaSeparatedPattern(allExcludes).mapToRegex()
         }
-        return validateConfig(config, loadDefaultConfig(), patterns())
+        return validateConfig(config, DefaultConfig.newInstance(), patterns())
     }
 }
