+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/config/InvalidConfig.kt
@@ -1,3 +1,3 @@
-package io.gitlab.arturbosch.detekt.cli.config
+package io.gitlab.arturbosch.detekt.core.config
 
 class InvalidConfig(override val message: String?) : RuntimeException(message, null, true, false)
