+++ detekt-core/src/main/resources/default-detekt-config.yml
