+++ detekt-core/src/test/resources/common_known_sections.yml
